requirements.txt
pywebview
qtpy
PyQt5
PyQtWebEngine
pytube